Create GMR_Magntics_and_plot_corners
To deploy the Ground Magnetic Resonance equipment the characteristics of the Magnetic field at that point and at that date need to be calculated. This software calculates these input parameters for places in Australia for the period 2010 to 2020.
The software calculates the GPS coordinates of the plot based in the azimuth of the magnetic field and the required size of the plot. Furthermore, the desired starting corner can be selected so that the plot extends away from the road or other obstacles. No need to estimate the center of the plot.